requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
contourpy==1.0.7
cycler==0.11.0
ffmpeg-python==0.2.0
filelock==3.9.0
fonttools==4.38.0
frozenlist==1.3.3
future==0.18.3
huggingface-hub==0.12.1
idna==3.4
kiwisolver==1.4.4
matplotlib==3.6.3
more-itertools==9.0.0
multidict==6.0.4
numpy==1.24.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
openai==0.26.5
openai-whisper==20230124
packaging==23.0
Pillow==9.4.0
pycparser==2.21
pyparsing==3.0.9
python-dateutil==2.8.2
PyYAML==6.0
regex==2022.10.31
requests==2.28.2
six==1.16.0
sounddevice==0.4.5
soundfile==0.11.0
tokenizers==0.13.2
torch==1.13.1
tqdm==4.64.1
transformers==4.26.1
typing_extensions==4.5.0
urllib3==1.26.14
yarl==1.8.2
